refactor: Print Student List component
